Anticipating exceptions
We used to have a workaround here, which would catch an exception caused by the file stream and then log it to Sentry immediately.
This was because if not logged an exception would be caught by the using statement of the GZipStream, which would then call dispose, which would in turn emit yet another exception.
By choosing to manually dispose of these streams, we're retaining all the functionality and we'll be able to catch the actual exceptions in some upper-level function.