Attempt repair content on failed content installation
Until now, when there was a CRC error while attempting to install
content, this error was non-recoverable.
Right now, we're allowing patcher to continue and attempting to fix
those files after unpacking is done.

Also, repairing process now checks if the files are okay after the
repair and tries again if not.